requirements.txt
whirlwind
whisper
tornado>=3.0
redis
tornado-redis
python-lust
